# Avoid repeated reading of the DeltaLog