--- a/core/src/main/scala/io/qbeast/core/model/RevisionClasses.scala
+++ b/core/src/main/scala/io/qbeast/core/model/RevisionClasses.scala
@@ -221,7 +221,7 @@ case class IndexStatus(
 case class CubeStatus(
     maxWeight: Weight,
     normalizedWeight: NormalizedWeight,
-    files: IISeq[QbeastFile])
+    files: IISeq[QbeastBlock])
     extends Serializable
 
 /**