--- a/src/main/scala/io/qbeast/spark/delta/IndexStatusBuilder.scala
+++ b/src/main/scala/io/qbeast/spark/delta/IndexStatusBuilder.scala
@@ -4,8 +4,7 @@
 package io.qbeast.spark.delta
 
 import io.qbeast.core.model._
-import io.qbeast.spark.utils.TagUtils
-import org.apache.spark.sql.{Dataset, SparkSession}
+import org.apache.spark.sql.Dataset
 
 import scala.collection.immutable.SortedMap
 
@@ -27,17 +26,9 @@ private[delta] class IndexStatusBuilder(
    * Dataset of files belonging to the specific revision
    * @return the dataset of AddFile actions
    */
-  def revisionFiles: Dataset[QbeastFile] = {
-    val spark = SparkSession.active
-    import spark.implicits._
-
+  def revisionFiles: Dataset[QbeastBlock] =
     // this must be external to the lambda, to avoid SerializationErrors
-    val revID = revision.revisionID.toString
-    qbeastSnapshot.snapshot.allFiles
-      .filter(_.tags(TagUtils.revision) == revID)
-      .map(addFile =>
-        QbeastFile(addFile.path, addFile.tags, addFile.size, addFile.modificationTime))
-  }
+    qbeastSnapshot.loadRevisionBlocks(revision.revisionID)
 
   def build(): IndexStatus = {
     IndexStatus(