… core/src/test/scala/io/qbeast/core/model/QbeastBlockTest.scala
… core/src/test/scala/io/qbeast/core/model/QbeastFileTest.scala
… src/main/scala/io/qbeast/spark/utils/Params.scala
… src/test/scala/io/qbeast/spark/QbeastIntegrationTestSpec.scala
… src/test/scala/io/qbeast/spark/delta/OTreeIndexTest.scala
… src/test/scala/io/qbeast/spark/delta/QbeastSnapshotTest.scala
… src/test/scala/io/qbeast/spark/index/CubeWeightsIntegrationTest.scala
… src/test/scala/io/qbeast/spark/index/OTreeAlgorithmTest.scala
… src/test/scala/io/qbeast/spark/index/query/QueryExecutorTest.scala