This PR fixes #61 

The changes made are:
- `OTreeIndex` now extends `FileIndex` instead of `TahoeLogFileIndex`
- QueryExecutor doesn't need `previouslyMatchedFiles` anymore
- `QbeastFile` now it's named `QbeastBlock`
- Add more metadata to the `QbeastBlock`: **size** and **modificationTime**
- `CubeStatus` now contains `QbeastBlock` objects instead of the path