@osopardo1 2022-01-26
https://github.com/Qbeast-io/qbeast-spark/pull/65#discussion_r792652606

In relation to this comment, I did the following comparisons using different approaches (collect before, collect after and the code in the main branch) in the `store_sales` table of the 10gb TPC-DS dataset. 

```scala
df.write
      .format("qbeast")
      .option("columnsToIndex", "ss_item_sk,ss_customer_sk")
      .option("cubeSize", "300000")
      .save(qbeastDir)
```

The times compared are from two different methods:
- **Query**. I compared running a count on the whole dataset both with Delta and Qbeast implementations
```scala 
val qbeast = spark.read.format("qbeast").load(qbeastDir)
val delta = spark.read.format("delta").load(qbeastDir)

qbeast.count()
delta.count()
```
- **ListFiles**. This is the method used by the TahoeLogFileIndex and OTreeIndex for filtering the files needed to query. It is useful to know if we improved something by reading only once the log
```scala
val deltaLog = DeltaLog.forTable(spark, qbeastDir)
val tahoeLogFileIndex =
  TahoeLogFileIndex(spark, deltaLog, deltaLog.dataPath, deltaLog.snapshot)
val oTreeIndex = OTreeIndex(tahoeLogFileIndex)

tahoeLogFileIndex.listFiles(partitionFilters = Seq.empty, dataFilters = Seq.empty)
oTreeIndex.listFiles(partitionFilters = Seq.empty, dataFilters = Seq.empty)
```
These are the implementations tested:

- Delta → qbeast indexed but read as delta
- Qbeast main branch → qbeast indexed but read with the main version
- Qbeast PR65 commit f2445ec1e58e61b23b5eea25d715b2a0c8831670 → Collect after `groupBy` in `buildCubeStatuses`
- Qbeast PR65 commit 36eea23a00be8f772df60ceb2dd86dd9cc8fec95 → Collect before `groupBy` in `buildCubeStatuses`

|Operation |Delta     |Qbeast Main       |QBeast PR65 Collect After  |Qbeast PR65 Collect Before         |
|------------|:----------:|:----------:|:----------:|:----------:|
|Query     |1.38    |2.94   |2.36   |1.73 |
|ListFiles | 0.18 | 1.4 | 0.9 |0.3 |

The tests were run locally. Feel free to discuss this approach here, and to suggest reversion of the changes! Thank you!