@osopardo1 2022-01-27
Made more test on a distributed cluster, just to have more numbers. Here are the results:

`mode`: DISTRIBUTED, `cluster`: 4 node cluster, `data`:10gb store sales, `columnsToIndex`: ss_customer_sk, ss_item_sk, `cubeSize` default , `num files`: 5000+ files

|Operation |Delta     |Qbeast Main       |QBeast PR65 Collect After  |Qbeast PR65 Collect Before         |
|------------|:----------:|:----------:|:----------:|:----------:|
|Query     |8.2s    |10.74s   |10.05s   |10.2s |
|ListFiles | 0.43s | 2.1s | 0.91s |0.53s |

`mode`: DISTRIBUTED, `cluster`: 4 node cluster, `data`:10gb store sales, `columnsToIndex`: ss_customer_sk, ss_item_sk, `cubeSize` 500000 , `num files`: 800+ files

|Operation |Delta     |Qbeast Main       |QBeast PR65 Collect After  |Qbeast PR65 Collect Before         |
|------------|:----------:|:----------:|:----------:|:----------:|
|Query     |1.64s    |2.19s   |3.14s   |2.55s |
|ListFiles | 0.32s | 0.65s | 0.96s |0.42s |


`mode`: DISTRIBUTED, `cluster`: 4 node cluster, `data`:100gb store sales, `columnsToIndex`: ss_customer_sk, ss_item_sk, `cubeSize` 50000000 , `num files`: 86 files

|Operation |Delta     |Qbeast Main       |QBeast PR65 Collect After  |Qbeast PR65 Collect Before         |
|------------|:----------:|:----------:|:----------:|:----------:|
|Query     |0.96s    |1.47s   |1.46s   |1.83s |
|ListFiles | 0.24s | 0.59s | 0.59s |0.42s |

We can see two things:
- Penalization of reading multiple files.
- For the largest DeltaLogs (for example the first one, with 5000+ files), this PR outputs better results on listFiles. In the other cases, the results are similar. 

I will revert to a `collect()` after the `groupBy` because due to https://github.com/delta-io/delta/issues/779 it an aggressive action with large DeltaLog can give problems. We could think about optimization later.